requirements.txt
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
idna==3.4
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
requests==2.28.2
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
